Nie mogę dalej tego ruszyć

+++ src/main/java/com/kodilla/testing/shape/Shape.java
@@ -4,5 +4,5 @@ public interface Shape {
 
     String getShapeName();
 
-    int getField();
+    double getField();
 }
